itsm/postman/urls.py
# -*- coding: utf-8 -*-
"""
Tencent is pleased to support the open source community by making BK-ITSM 蓝鲸流程服务 available.

Copyright (C) 2021 THL A29 Limited, a Tencent company.  All rights reserved.

BK-ITSM 蓝鲸流程服务 is licensed under the MIT License.

License for BK-ITSM 蓝鲸流程服务:
--------------------------------------------------------------------
Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated
documentation files (the "Software"), to deal in the Software without restriction, including without limitation
the rights to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the Software,
and to permit persons to whom the Software is furnished to do so, subject to the following conditions:

The above copyright notice and this permission notice shall be included in all copies or substantial
portions of the Software.

THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT
LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN
NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY,
WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE
SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
"""

from django.conf.urls import url
from rest_framework.routers import DefaultRouter

from itsm.postman.views import (
    ApiInstanceViewsSet,
    RemoteApiViewSet,
    RemoteSystemViewSet,
    RpcApiViewSet,
)


routers = DefaultRouter(trailing_slash=True)

routers.register(r'api_instance', ApiInstanceViewsSet, basename="api_instance")

routers.register(r'remote_system', RemoteSystemViewSet, basename='remote_system')

routers.register(r'remote_api', RemoteApiViewSet, basename='remote_api')

# APIView不能通过routers.register()的方式注入路由
urlpatterns = routers.urls + [
    url(r'^rpc_api/$', RpcApiViewSet.as_view()),
]
